# PERF: passthrough mask to take_1d if already known