--- a/pandas/core/internals/array_manager.py
+++ b/pandas/core/internals/array_manager.py
@@ -48,6 +48,7 @@ from pandas.core.dtypes.generic import (
 )
 from pandas.core.dtypes.inference import is_inferred_bool_dtype
 from pandas.core.dtypes.missing import (
+    _isna_array,
     array_equals,
     isna,
     na_value_for_dtype,
@@ -610,16 +611,15 @@ class BaseArrayManager(DataManager):
         else:
             validate_indices(indexer, len(self._axes[0]))
             indexer = ensure_platform_int(indexer)
-            if (indexer == -1).any():
-                allow_fill = True
-            else:
-                allow_fill = False
+            mask = indexer == -1
+            needs_masking = mask.any()
             new_arrays = [
                 take_1d(
                     arr,
                     indexer,
-                    allow_fill=allow_fill,
+                    allow_fill=needs_masking,
                     fill_value=fill_value,
+                    mask=mask,
                     # if fill_value is not None else blk.fill_value
                 )
                 for arr in self.arrays
@@ -1056,22 +1056,32 @@ class ArrayManager(BaseArrayManager):
         if unstacker.mask.all():
             new_indexer = indexer
             allow_fill = False
+            new_mask2D = None
+            needs_masking = None
         else:
             new_indexer = np.full(unstacker.mask.shape, -1)
             new_indexer[unstacker.mask] = indexer
             allow_fill = True
+            # calculating the full mask once and passing it to take_1d is faster
+            # than letting take_1d calculate it in each repeated call
+            new_mask2D = (~unstacker.mask).reshape(*unstacker.full_shape)
+            needs_masking = new_mask2D.any(axis=0)
         new_indexer2D = new_indexer.reshape(*unstacker.full_shape)
         new_indexer2D = ensure_platform_int(new_indexer2D)
 
         new_arrays = []
         for arr in self.arrays:
             for i in range(unstacker.full_shape[1]):
-                new_arr = take_1d(
-                    arr,
-                    new_indexer2D[:, i],
-                    allow_fill=allow_fill,
-                    fill_value=fill_value,
-                )
+                if allow_fill:
+                    new_arr = take_1d(
+                        arr,
+                        new_indexer2D[:, i],
+                        allow_fill=needs_masking[i],
+                        fill_value=fill_value,
+                        mask=new_mask2D[:, i],
+                    )
+                else:
+                    new_arr = take_1d(arr, new_indexer2D[:, i], allow_fill=False)
                 new_arrays.append(new_arr)
 
         new_index = unstacker.new_index