@jreback 2021-11-29
lgtm cc @jbrockmendel 

maybe more cases that could use this (for followon for sure)