@jbrockmendel 2021-11-29
nice perf bump.  any interest in applying it to BM while you're at it