@jorisvandenbossche 2021-11-29
>  any interest in applying it to BM while you're at it

I don't think the same optimization is directly applicable to BlockManager: 1) the one place where the Block uses `take_nd` directly in `reindex_indexer`, the mask is not precomputed (and it also wouldn't help to pre-compute, since computing it inside take_nd is as efficient, since this is done on the full 2D array), and 2) `unstack` doesn't use `take` (and in general wouldn't do repeated `take` calls with the same indexer, I think, since it is block-based).